utils: move get_pb_type_chain funcion to lib/pb_type
